Separate cli code from main binary entrypoints, move to cmd folder
