+++ cmd/atc/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/facebook/augmented-traffic-control/cmd/atc/cli"
+
+func main() {
+	cli.Execute()
+}
