+++ cmd/atc_api/cli/cli_test.go
@@ -1,4 +1,4 @@
-package main
+package cli
 
 import (
 	"os"
