+++ cmd/atc_api/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/facebook/augmented-traffic-control/cmd/atc_api/cli"
+)
+
+func main() {
+	cli.Execute()
+}
