+++ cmd/atcd/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/facebook/augmented-traffic-control/cmd/atcd/cli"
+
+func main() {
+	cli.Execute()
+}
